Squashed commit of the following
commit 494549d5 
Author: bdamja <iceicebabyalter@gmail.com> 
Date: Mon Jan 31 2022 14:17:06 GMT-0800 (Pacific Standard Time) 

    fixed formatting


commit db7205af 
Author: bdamja <iceicebabyalter@gmail.com> 
Date: Mon Jan 31 2022 14:05:31 GMT-0800 (Pacific Standard Time) 

    created a toTicks method, takes float seconds and returns int ticks

+++ BridgeTeams/src/main/java/gg/bridgesyndicate/bridgeteams/GameDataPoller.java
@@ -1,6 +1,6 @@
 package gg.bridgesyndicate.bridgeteams;
 
 public interface GameDataPoller {
-    int NO_START_ABORT_TIME_IN_SECONDS = 120;
+    float NO_START_ABORT_TIME_IN_SECONDS = 120.0f;
     void poll(BridgeTeams bridgeTeams);
 }
